moved symmetry to types.py
